This commit was manufactured by cvs2svn to create tag 'IGB_RELEASE_3_24'.
git-svn-id: http://svn.code.sf.net/p/genoviz/code/tags/IGB_RELEASE_3_24@239 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd

+++ igb/src/com/affymetrix/igb/prefs/GraphsView.java
@@ -66,7 +66,7 @@ public class GraphsView extends JPanel implements IPrefEditorComponent  {
 
     JLabel graph_height_label2 = new JLabel("Tiered graph default coord height: ");
     JTextField graph_height_field2 = UnibrowPrefsUtil.createNumberTextField(
-      GraphGlyphUtils.getGraphPrefsNode(), GraphGlyphUtils.PREF_ATTACHED_COORD_HEIGHT, Double.toString(GraphGlyphUtils.default_coord_height), Double.class);
+      GraphGlyphUtils.getGraphPrefsNode(), GraphGlyphUtils.PREF_ATTACHED_COORD_HEIGHT, Integer.toString(GraphGlyphUtils.default_coord_height), Integer.class);
     Box height_box2 = new Box(BoxLayout.Y_AXIS);
     height_box2.add(graph_height_label2);
     height_box2.add(Box.createHorizontalStrut(5));
